+++ src/feature/my/component/index.js
@@ -1,0 +1,6 @@
+export { default as AccountTab } from './AccountTab/AccountTab';
+export { default as ActivityTab } from './ActivityTab/ActivityTab';
+export { default as CircleProfile } from './CircleProfile/CircleProfile';
+export { default as MyInfo } from './MyInfo/MyInfo';
+export { default as PolicyTab } from './PolicyTab/PolicyTab';
+export { default as TopOverlay } from './TopOverlay/TopOverlay';
