+++ src/hooks/index.js
@@ -3,4 +3,3 @@ export { default as useLike } from './useLike';
 export { default as usePopUp } from './usePopUp.jsx';
 export { default as useScrap } from './useScrap';
 export { default as useSearch } from './useSearch';
-export { default as useTabView } from './useTabView';
