+++ src/pages/MyPage/index.js
@@ -7,10 +7,3 @@ export { default as ChangePasswordPage } from './pages/ChangePasswordPage/Change
 export { default as DeleteAccountPage } from './pages/DeleteAccountPage/DeleteAccountPage';
 export { default as PrivacyPolicyPage } from './pages/PolicyPage/PrivacyPolicyPage';
 export { default as ServicePolicyPage } from './pages/PolicyPage/ServicePolicyPage';
-
-export { default as TopOverlay } from './components/TopOverlay/TopOverlay';
-export { default as CircleProfile } from './components/CircleProfile/CircleProfile';
-export { default as MyInfo } from './components/MyInfo/MyInfo';
-export { default as AccountTab } from './components/AccountTab/AccountTab';
-export { default as ActivityTab } from './components/ActivityTab/ActivityTab';
-export { default as PolicyTab } from './components/PolicyTab/PolicyTab';
